+++ includes/blocks.php
@@ -113,10 +113,16 @@ add_action( 'init', function() {
  * add_action or add_filter features (rather than adding
  * these hooks outside the block's directory within functions.php).
  */
-$blocks = glob( AQUAMIN_ASSETS . '/block-library/*/hooks.php' );
-if ( $blocks ) {
-	foreach ( $blocks as $block ) {
-		require_once $block;
+$hook_paths = array(
+	AQUAMIN_ASSETS . '/block-library/*/hooks.php',
+	AQUAMIN_ASSETS . '/block-edits/*/hooks.php'
+);
+foreach ( $hook_paths as $path ) {
+	$hooks = glob( $path );
+	if ( $hooks ) {
+		foreach ( $hooks as $hooks ) {
+			require_once $hooks;
+		}
 	}
 }
 
@@ -129,7 +135,7 @@ add_filter( 'block_categories' , function( $categories ) {
 } );
 
 /**
- * Manage block patterns and block pattern categories
+ * Manage Expose all synced patterns in the WordPress back-end sidebar
  *
  * @see https://developer.wordpress.org/block-editor/reference-guides/block-api/block-patterns/
  */
@@ -148,122 +154,21 @@ add_action( 'after_setup_theme', function() {
 		);
 	}, 10, 2 );
 
-	// create the patterns post type
-	register_post_type(
-		'custom-patterns',
-		array(
-			'labels' => array(
-				'name' => _x( 'Block Patterns', 'Taxonomy General Name', 'aquamin' ),
-				'singular_name' => _x( 'Block Pattern', 'Taxonopmy Singular Name', 'aquamin' ),
-				'menu_name' => __( 'Block Patterns', 'aquamin' ),
-				'all_items' => __( 'All Block Patterns', 'aquamin' ),
-				'parent_item' => __( 'Parent Block Pattern', 'aquamin' ),
-				'parent_item_colon' => __( 'Parent Block Pattern:', 'aquamin' ),
-				'new_item_name' => __( 'New Block Pattern Name', 'aquamin' ),
-				'add_new_item' => __( 'Add New Block Pattern', 'aquamin' ),
-				'edit_item' => __( 'Edit Block Pattern', 'aquamin' ),
-				'update_item' => __( 'Update Block Pattern', 'aquamin' ),
-				'separate_items_with_commas' => __( 'Separate Block Patterns with commas', 'aquamin' ),
-				'search_items' => __( 'Search Block Patterns', 'aquamin' ),
-				'add_or_remove_items' => __( 'Add or remove items', 'aquamin' ),
-				'choose_from_most_used' => __( 'Choose from the most used items', 'aquamin' ),
-				'not_found' => __( 'Not Found', 'aquamin' ),
-			),
-			'public' => true,
-			'publicly_queryable' => is_user_logged_in(),
-			'menu_position' => 20,
-			'menu_icon' => 'dashicons-layout',
-			'show_in_rest' => true,
-			'supports' => array(
-				'editor',
-				'custom-fields',
-				'title',
-				'thumbnail',
-			),
-		)
-	);
-
-	// create patterns categories
-	register_taxonomy(
-		'custom-patterns-categories',
-		'custom-patterns',
-		array(
-			'show_ui' => true,
-			'show_admin_column'	=> true,
-			'public' => true,
-			'publicly_queryable' => is_user_logged_in(),
-			'hierarchical' => false,
-			'show_in_rest' => true,
-			'labels' => array(
-				'name' => 'Pattern Categories', 'Taxonomy General Name', 'aquamin',
-				'singular_name' => 'Pattern Category', 'Taxonomy Singular Name', 'aquamin',
-				'menu_name' => 'Pattern Categories', 'aquamin',
-				'all_items' => 'All Items', 'aquamin',
-				'parent_item' => 'Parent Item', 'aquamin',
-				'parent_item_colon' => 'Parent Item:', 'aquamin',
-				'new_item_name' => 'New Item Name', 'aquamin',
-				'add_new_item' => 'Add New Item', 'aquamin',
-				'edit_item' => 'Edit Item', 'aquamin',
-				'update_item' => 'Update Item', 'aquamin',
-				'separate_items_with_commas' => 'Separate items with commas', 'aquamin',
-				'search_items' => 'Search Items', 'aquamin',
-				'add_or_remove_items' => 'Add or remove items', 'aquamin',
-				'choose_from_most_used' => 'Choose from the most used items', 'aquamin',
-				'not_found' => 'Not Found', 'aquamin',
-			),
-			'capabilities' => array(
-				'manage__terms' => 'edit_posts',
-				'edit_terms' => 'manage_categories',
-				'delete_terms' => 'manage_categories',
-				'assign_terms' => 'edit_posts'
-			)
-		)
-	);
-
-	// register pattern categories in the block editor
-	$categories = get_terms( 'custom-patterns-categories' );
-	if ( $categories ) {
-		foreach( $categories as $cat ) {
-			register_block_pattern_category(
-				$cat->slug,
-				array( 'label' => $cat->name )
-			);
-		}
-	}
-
-	// register patterns in the block editor
-	$patterns = get_posts( array(
-		'post_type' => 'custom-patterns',
-		'posts_per_page' => 999,
-	) );
-
-	if ( $patterns ) {
-		foreach( $patterns as $pattern ) {
-			$cats = array( 'uncategorized' );
-			$applied_cats = wp_get_post_terms( $pattern->ID, 'custom-patterns-categories' );
-			if ( $applied_cats ) {
-				$cats = array_map(function ($cat) { return $cat->slug; }, $applied_cats);
-			}
-			register_block_pattern(
-				'aquamin/pattern-' . $pattern->post_name,
-				array(
-					'categories'  => $cats,
-					'title'       => $pattern->post_title,
-					'content'     => $pattern->post_content,
-				)
-			);
-		}
-	}
-
 } );
 
 /**
- * Expose all reusable blocks in the WordPress back-end sidebar
+ * Expose all synced patterns under Appearance > Synced Patterns
  * 
  * @link https://www.billerickson.net/reusable-blocks-accessible-in-wordpress-admin-area
  */
 add_action( 'admin_menu', function() {
-    add_menu_page( __( 'Reusable Blocks', 'aquamin' ), __( 'Reusable Blocks', 'aquamin' ), 'edit_posts', 'edit.php?post_type=wp_block', '', 'dashicons-editor-table', 22 );
+	add_submenu_page(
+        'themes.php',
+        __( 'Synced Patterns', 'aquamin' ),
+        __( 'Synced Patterns', 'aquamin' ),
+        'edit_posts',
+        'edit.php?post_type=wp_block'
+    );
 } );
 
 /**