+++ docs/changelog.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## 6.2.0 - November 29 2023
-
+* Remove custom block patterns feature (this is now supported by core WordPress).
+* Move synced pattern to Appearance > Synced Patterns to clean up the sidebar.
 
 ## 6.1.1 - October 5 2023
 * Fix global padding and block gap position in theme.json so they apply properly.